requirements.txt
Flask[async]==2.2.0
Werkzeug==2.2.2 # otherwise import error for flask
gunicorn==20.1.0
google-cloud-storage==1.43.0
# oauth2client==4.1.3
google-cloud-logging==2.7.0
google-cloud-tasks==2.7.1
google-api-python-client==2.34.0
# google-auth==2.17.3
google-api-core==2.12.0
google-cloud-bigquery==3.25.0
numpy==1.26.4
#yt-dlp==2025.1.12
# yt-dlp
pandas_gbq==0.28.0
pandas==2.2.2
# vertexai==1.108.0
ffmpeg-python==0.2.0
moviepy==1.0.3 